feat: Integración base datos

+++ app_movile/lib/features/home/data/model/trip_statistics.dart
@@ -1,5 +1,5 @@
 class TripStatistics {
-  final int weeklyDistance;
+  final double weeklyDistance;
   final int calories;
   final int goalPercentage;
   final bool isDeviceConnected;
